Minor improvements
* Couple random items for Evelynn Rose to hand out instead of her
fucking gun, along with a backpack to ensure she can carry it.
* Allow breaking apart bio-weapon badges into silver.
* Downgraded Alphas's Apex Predator to Hunter, as the intelligence
penalty has weird effects in character generation.
* Removed Chaos threshold for UI reasons, after confirming it still
permits starting with Eater of The Dead without it. Leaving the fungal
failed bio-weapon with Mycus is both fine UI-wise and vital, as it makes
sense for Mycus to override that, plus it needs the threshold to
function.
* Replaced Gamma's Very Smart mutation with Bookworm, as the int bonus
likewise has weird effects if you switch professions during chargen.