Assorted changes
* Fixed bio-weapon badges turning into nothing when uncrafted. Turns out
that that feature doesn't work despite it being used, and for all I know
it may have never worked.
* Some changes to itemgroup spawns in the makeshift command center, as
varible-repeat is now WAY less predictable. Oughta fix up the other
locations sometime.
* Replaced the command center's indoor charcoal smoker with a pair
outdoors, as using it for non-recipe smoking indoors is a Bad Idea.
* Moved that one guard a step down so they aren't in the middle of the
walking path.
* Rebalanced and reworded the laser gatling gun, because sadly "takes up
an arm" is hardcoded into the fusion blaster arm. Main effects are
reducing cost to 50 and burst size to 50, sae as fusion blaster (which
has more damag). Power per activation is consumed on a per-shot basics
when full-auto is involved, so that's a burst of laser shots for 250
power vs "dump 2000 power and one target dies instantly"