Second commit
1. Added backstory system for Evelynn Rose when following the player.
Dynamic lines used change depending on if the player has successfully
convinced her to reveal her military background, and include lines that
let the player a chance to do so if they hadn't already. Convincing her
from this dialogue section is intentionally easier than it was when she
was still a stranger, without the same risk of making her distrust you.
2. Made it so that the persuasion option for Evelynn requires slightly
lower stats than before, and allows perception as an alternative to
intelligence. Also implemented the use of an effect so that the
conversation option is only available if you haven't already convinced
her to open up. Also includes a 3-hour cooldown on pestering her about
it if failed.
3. Minor updates to Evelynn's initial dialogue, mostly grammar and word
usage in a couple places.
4. Removed the generic bio-weapon NPC class since it's no longer used by
Sigma and Lambda.
5. Gave the BG_survival_story_RURAL trait group to prepper NPCs.
6. Expanded the barebones dialogue of the Best Friend a bit, by making
their dialogue resemble that of a mundane neutral survivor if the player
has certain traits that imply not having a normal background. This
includes churls in vanilla, and the bioweapons for Cata++.
7. Changed part of Router's dialogue to change in the event that the
player has most likely killed off Apophis.
8. Reworked recruiting Sigma and Lambda to involve being on the mission
to hunt Apophis.
9. Added a touch of added dialogue for Router, Sigma, and Lambda. After
completing the mission it becomes possible to ask for advice from
Router, opening up a path to potentially recruit Sigma and Lambda if
they weren't recruited during the mission.