Update battery items and weaponry
* Overhauled the fancy laser weapons once again. Now they use atomic
batteries instead of UPS, as a properly-function version of the older
"use plutonium cells outright" idea Noct had.
* Also rebalanced the Omnitech laser weapons to use variously FIRE_20 (I
only just now learned that's a thing), FIRE_50, or FIRE_100. In addition
they make use of reload time now (generally slower than changing a mag
for mundane guns), with description tweaks to mention the change, and
removal of some properties that were just setting the default 0.
* Set the NPCs that had UPS weapons to now use the right sorts of
atomic-powered versions of the weapon, and have the right sorts of cells
on hand.
* Because it turns out that neither the atomic batteries nor their
recipe book SPAWN anywhere in vanilla, some itemgroup entries have been
added to allow them to spawn in likely places, to make them rare but not
impossible to find.
* Changed zombie super soldiers to drop their weapons with random
amounts of charge left in the battery, and a chance of dropping a
less-depleted spare battery.
* Added disposable batteries to the options available for loading
survivor battery guns.