Update survivor's crank rifle, other laser fixes
* Overhauled the survivor's crank rifle. It uses a homemade battery
bank, which is implied to equal a total capacity of 300 kJ via its added
recipe. This is to make up for the need to load a battery into it to use
the `HAND_CRANK` action, using a propretary battery when it only uses 1
charge at a time makes its abnormal degree of power amplification stand
out less, and excludes the ability to charge normal batteries with it.
* Updated the recipe as well, making it so that it uses a hand-crank
charger instead of combining a foot crank with a car alternator, but
also made the misc electronics more distinct for flavor, with power
amplifiers being the most obvious addition.
* Added a magazine for the crank rifle and cooresponding recipe, stated
to be as much a capacitor bank as it is a bunch of batteries.
* Removed decomp learn from the crank rifle recipe, since it can't
actually be dissaembled.
* Removed mispelled two-handed use flags from a few weapons where they
didn't make sense, replaced with `FIRE_TWOHAND` for the one case where
it remained relevent to its function.
* Fixed Cata++ lasers fouling up via liberal application of the
`NON_FOULING` flag.