Tweak to military suits, survivor plate armor
* Shifted the encumberance ranges of military armor. The light version
is now 5 points less encumbering, the heavy variant 5 points more. Not
yet sure if it's actually intended that the military suits be impossible
to refit, but this pushes them into interesting niches compared to
standard survivor suits.
* Implements an idea that came to mind when noticing that there's a
modified mail armor option now, but nothing for other medieval armor.
Basically the bastard lovechild of plate armor and a survivor suit. Plus
this fills another niche between survivor suits and heavy survivor
suits, for which there's a LOT of room.
* Couple typo fixes along the way.