Updates to regen, harvest, recipes
* Implemented the new variable regen instead of the now-unused
regeneration flags. On the plus side, this means now the super soldiers
and bio-weapons can all get varying levels of regen.
* Updated harvest entries so that bio-weapons make use of mutant
humanoid meat and fat, along with large human stomachs (forgot that was
a thing).
* Removed overrides for mutagen as mutagen crafting has been extensively
updated since then, and it was now causing load errors.
* Changed the welding goggles recipe to an alternative version using
suffix, so it's no longer a full override.