Fleshing out Router's missions: part one
* Set it up so that Router is skepitical of players who don't have a
bio-weapon badge on hand, allowing the potential for a side mission that
lets players find the bio-weapon lab if they've reached the command
center first.
* Players who started Evy's mission first can opt to skip this, as
that's the only way for the game to easily track that the player already
has visited the bio-weapon lab. Sadly, it's not available to players who
recruited Evy in a save predating this PR, but it's a start at least.
* Bio-weapons and super soldiers likewise outright skip the side
mission, as does any player who already has a bio-weapon badge on hand.