Updates to Super Soldier scenario usage
* Overhauled the Experiemental Soldier Start. Set it to instead start
you outside one of several lab locations, generally the ones that're
unique or at least garaunteed to soawn. The default is now the Research
Facility, with Hub 01, the outside of a Bio-Weapon Lab, the Makeshift
Command Center, and the Hazardous Waste Sarcophagus as alternatives.
This allows for alternative ways to start up the Bio-Weapon mission
chain.
* Added the super soldiers as valid spawns for the Lab Staff challenge,
preserving the original idea behind the Experimental Soldier Start in a
way that makes use of an existing scenario.
* Added the super soldiers to Ambushed, Helicopter Crash and The Next
Summer starts, as these are perhaps the most logical scenarios to have
an advanced super soldier show up.
* Set it so that the super soldiers have a cost in the range of 8-10
points, which is the norm for most advanced professions (and puts them
more in line with their advantages over the bionic soldier professions).
This also obviates having to tack on a 4-point bonus for Experimental
Soldier Start.
* Removed some starting locations that aren't actually used, done while
adding the new ones.