Laser-related updates
* Beefed up the min and max range of SMG super soldiers to 3-15.
* Updated the monster overrides for super soldier lasers, removing
INCENDIARY effect because it seems to not behave right anymore when used
on players, plus fixed their ammo not acting like player-fired variants
due to monster guns not respecting FIRE_X flags.
* Also updated monster ranged weapon dispersion, generally now
consistent with the monster's max attack range. This is a buff in the
case of pistol bio-engineers, but reduces accuracy for other types.
* Removed overrides for vanilla lasers since there were some issues with
that (like causing electrolaser gunmods to give the V29 negative range),
with the Omnitech weapons rebalanced accordingly in the few places it
was needed. In general it seems reasonable for Omnitech lasers to be
more damaging per UPS consumed (80% damage/power, whereas vanilla UPS
guns have around 50% rating), have superior accuracy due to actually
getting the "low dispersion" laser thing, but have relatively lower
range and worse piercing.
* Increased the NEO-33's range to 16 and its pierce to 5.
* Reduced the AKRO-388's damage to 16 as well, since it has same power
draw as the NEO-33.
* Lowered the ARC-314's damage to 40, reduced range to 28.
* Reduced the KRX-108's damage to 40 (consistent with ARC-314), reduced
range to 40.
* Reduced the MX-84's damage to 80, once again putting it in the samage
damage/power range as the other weapons.
* Inceased the XARM-37's damage to 40, main drawbacks are still having
short range and having half as much pierce as the rifle and LMG.
* Couple description typofixes along the way.