Overhaul Biological Weapons: Part One
* Added custom adrenaline effect for the biological weapons to make use
of. While it has the advantage of also stopping bleeding, its effective
duration is much shorter than its cooldown (int_dur_factor of 180 means
positives start at 3 minutes duration, max_duration of 200 means you'll
only get 20 seconds of actual adrenaline), requiring continued use to
stave off the comedown effects.
* Implemented custom spell-like effects that the biological weapons will
make use of. The outermost effect applies bleeding to a target (will be
used by active version of the sword), while beneath that are effects
that have a one-third chance of granting the custom adrenaline rush, for
enough seconds to continue using it with repeated effort. The adrenaline
effect is planned to be used on target-hit for the sword, and called
when the player is hit for the biological guns.
* Shuffled around the AEP effects for biological knife and sword.
General trend is for less severe effects when inactive (AEP_SICK when
carried, AEP_SCHIZO and AEP_SAP_LIFE when wielded), and removed ability
to call AEA effects from the inactive version of the sword. Also removed
artifact use action from the biological knife, as it's not needed for an
item with no AEAs.
* Active sword still retains its "+2 str and dex, -2 int and per, +20
speed" effect, along with gaining AEP_EVIL and AEP_MUTAGENIC, but now
also gains AEP_THIRST and AEP_HUNGER.
* Nerfed the cutting damage of active biological sword, halving it to 50
(which is double that of the inactive version).
* Added the "needs to be wielded" property to the biological sword's
transformation.
* Set it so that activating the sword uses its actual charge level,
instead of automatically setting it to 10.
* Removed NON_STUCK flag since weapon sticking has been gone for a long
time, and the flag is no longer used.
* Enabled sheathing the biological knife and (inactive) sword.
* Made sure traders won't want to try and buy the active biological
sword, via TRADER_AVOID flag.
* Also gave the biological melee weapons MAGIC_FOCUS for flavor.
* Added weapon category to the converted melee CBMs and to the
biological knife/sword, the latter fixing their counting as artifacts.
* Couple description updates and typofixes along the way.
* Set the active version of the sword to apply the spell features, on
hitting the target. Hitting an enemy with it inflicts blood loss, while
granting the custom adrenaline effect to the user. This is currently not
functional due to a bug, but will be up and running as soon as the bug
reposible is fixed in CDDA proper.
* Also removed tool qualities from the active version as built-in use
actions don't play nice with weapons that rely on revert instead of
transform.