Plasma Pulse Cell Update
* Added proper compressed hydrogen gas, and an energy-intensive recipe
to produce it. Useful if you find a need for hydrogen, but not practical
since hydrogen takes a fair bit of power to produce (especially since
post-cata electrolysis is little more than two electrodes and a
battery).
* Added an override changing vanilla hydrogen tanks from a tool into a
magazine, filled with hydrogen gas. These two changes are used for...
* The Plasma Pulse Cell CBM, designed to draw power from hydrogen gas,
with hydrogen canisters as an even beefier alternative. Even at 25%
efficiency, due to the low stack size of these items plus the high
energy density of hydrogen, each unit of either is a massive spike of
power, and enough waste heat dumped into the air that you pretty much
need to keep moving if you're using it a ton.
* Replaced Beta's use of slow JTR with plasma pulse cell CBM, which is
much more suitable for the most power-hungry of the Bio-Weapons.
* Fixed super juggernaut profession having no powergen CBM via giving
them the gasoline fuel cell CBM.
* Also removed the RTG inductor from super juggernauts, and from the
droplist of their undead counterparts.
* Updated the harvest results of augment undead soldiers to reflect
updates to their profession version's CBM list.
* Added some powergen CBMs to the harvest for augmented abominations, to
imply failed bio-weapons are more variable in augmentations compared to
the playable version, in line with the MI+JTR combo being intended as a
mix of two other Bio-Weapon's standard kit (which will be the case if I
can ever devise something unique for Delta).
* Similarly, updated the CBMs drops for generic zombie super-soldiers,
based on current overlapping CBMs and including all of the powergen
options common to the four (which will be specific to each eventually,
since I plan to give the Scout something that will be specific to
Delta).
* Gave failed bio-weapons a chance of dropping a random collection of
fuel/ammo/batteries, based on what their non-failure versions would
regard as fuel.
* Organized non-power-storage CBMs in the super soldier profession
bionics alphabetically, to make it easier to compare with the monster
drop lists.
* Misc updates to description, typofixes along the way, also updated
starting power of super soldier juggernauts to be consistent with other
supers soldiers, and gave Scout the Nanites+Painkiller combo all over
super soldiers have.