Atomic Battery CBM
Added an Atomic Battery CBM. This was originally going to be flavored as
an RTG, but the main issue is that "perpetual power" works a bit
backwards if not using one of the seemingly-hardcoded power sources. It
generates power passively when off, but can't be set to generate power
properly when turned on. It also can't be made truly passive since it's
a powergen bionic, trying to disable the ability to toggle it just left
it so you could still toggle it. Once, permanently disabling it.

Hence, the "disconnent switch" mechanism is flavored to disguise the
fact that it has to work the exact opposite of how powergen CBMs
normally work. This also precluded doing anything neat like having it
give off heat or radiation when active. So instead, after already having
thought up a rough idea for how it works, I realized it made more sense
as an alphavolatic cell.

Power generated is very low, 18 kJ per hour, but it's also completely
free, and builds up 24/7 unless you deliberately turn it off. I was
going to go with a larger figure, enough to make it useful as a primary
power source, but at that rate it made the use of long-action CBMs like
Integrated Toolest completely free due to charging faster than it could
deplete. This seems about right for purely-secondary powergen rather
than a primary, enough that downtime spent crafting or reading will
grant usable reserves, without being easily abused.

Added thm to appropriate spawns, potential failed bio-weapon harvest,
etc. Also gave it to Bio-Weapon Delta, as their seconary powergen
source, displacing Metabolic Interchange to make Joint Torsion Ratchet
their primary power source. This better suits them, and finally
accomplishes my intended objective of making each bio-weapon's power
generation setup unique, with both a primary and secondary.

Super Scout meanwhile uses JTR as their primary, since Atomic Battery
would be too low-yield to be practical as primary powergen. This
accomplishes my other objective, of making the super soldiers each use
the same primary powergen as one of the main bio-weapons (assuming that
Gasoline Fuel Cell is Beta's primary, which makes sense since Plasma
Pulse Cell is more of an emergency-use CBM).