Changes/fixes to scenario mutation restrictions
* Reduced the number of mutations barred from Bio-Weapon Lab and. Now
that telling it to bar a trait that a profession HAS causes errors and
removes that profession from the scenario entirely, the more extensive
restiction was causing the scenarios to break entirely.
* Instead went with only blacklisting traits that logically would
pertain to growing up and having normal experiences. This of course
means a lot more traits can be selected, but still avoids any trait that
seems especially abnormal for a lab-raised supeweapon, as best as their
wide variety of forced mutations allows.
* Updated that trait list with some of the more recent traits, many of
which are odd flavor things that fall in the category covered by point
two.