Various gunmod/item updates
* Overhauled the auxillery pistols to use a mix of stats based on a
selection of pistols, and enable using a wider range of magazines. This
comes at the expense of accuracy, reliablity, and in the case of 9mm it
can't use the 50 and 100-round Glock mags. Allowing the survivor SMG
mags however seemed like an interesting addition to add however, getting
more potential use of those items.
* Recipes for the auxillery pistols involving a bit of work modifiyng
one of several relatively common pistols as a component.
* Updated the LMG handle's values to focus more on aim speed and recoil
reduction, with its main penalites stacked onto dispersion and sight
dispersion. Reducing the effect of felt recoil seemed closer to logical
compared to it being better at sight dispersion than iron sights.
* Made the LMG carry handle no longer integral to the LMGs, plus removed
use of it from the pneumatic LMG outright since it requires your two
hands to fire. Made it craftable and attachable as a separate item
instead.
* Belatedly updated recipes so that drilling will demand level 2, as
level 1 is unused in vanilla but used in mods for non-metal drills.
* Made a few other Survivor;s Cookbook allow alternative books to learn
them from later on, since a good chunk of the items already did that.
* Updated the makeshift receiver items to use the specified casing as a
tool instead of a component.
* Reworked the full-auto case and receiver items. Basic gist of it is
that the old autofire mod is obsolete and the closest sear mods require
a non-makeshift gun to acquire. So, bumped up the skill demands of the
auto case accordingly, altering its flavor to be more like a
customizable equivalent to the drop-in sear item (with additional parts
to faciitate building a custom receiver).
* Altered itemgroups a fair bit. Makeshift full auto parts got removed
from military-focused itemgroups but added to some where their inclusion
makes sense (including drug dealers), some no-longer-used itemgroups
were removed, proper gunmod itemgroup was used as well. Makeshift guns
also get weights that better account for the sorts of weighting used in
the group they appear in.
* Also minor stat updates to survivor's laser sight.
* Slight tweak to a stat in the aux flamethrower I missed.