Metal molding updates, and molded crowbars
* Updated use of metal molding requirement to only specify the mold
items and hammering quality, allowing the use of steel_standard or
steel_tiny as needed. And crafting mechanics mean that max charges in
the can forge no longer actually blocks crafting, it merely requires a
couple reloads.
* Reworked bearing recipe with the latest update to bearing crafting in
mind.
* Reworked time for a few molding recipes. Basically the gist is 30
minutes or 10 minutes per lump of steel used, whichever is higher (the
latter seems to be how time was already deterimed, just the former adds
a minimum). In exchange, batch time savings are now implemented.
* Added metal-molding recipes for rebar, nails, and blades. Blade recipe
is the odd one old, needing cutting quality instead of specific items as
the mold source, as it was hard to decide what exact item to base the
mold off of. Whereas bearings, rebar, and nails are all "have one, need
more" scenarios, blades are in the same boat as the molded tools, in
that if you have one already than there is no need for the recipe, which
is a bit of a problem for blade items normally (case in point, can opt
to break down a servicable machete or broadsword for something you only
need if you're so hard up for weapons that you'd be willing to make a
makeshift machete).
* Also added a molded crowbar, that lets you gain the quality needed to
actually handle doors using a makeshift crowbar as the tool needed by
the mold.
* Finally, belatedly reworked the molded hammer since they eventually
removed prying ability from the hammers you use as a mold source.
* Added SHEATH_AXE to molded axe, replaced SHEATH_KNIFE with BELT_CLIP
for molded hammer.