Update corpse spawns
* Used `remains_soldier` instead of the now-removed `map_extra_military`
itemgroup. Nice and fitting for the Unknown Lab, since that leaves
gibbed bodyparts instead of corpses.
* Gave the Survivor Holdout's corpses a bit more variety, including a
few other variants of the "nothing but remains left" itemgroups, since
only one instance of `map_extra_military` was called here, and it'd look
weird if the one soldier among the bodies was the only non-intact
corpse.