Branch set aside for pre-nested containers release
Do not merge this into the base repo, as this is technically a
regression.

* Monster armor values updated, removing ballistic resistance as it
post-dates nested containers. Use of stab armor of equivalent value
instead, as this was used by ballistic weapons at the time.
* Conversion of ballistic weapons to deal stab damage, for reasons as
stated above. Weapon overrides for the vanilla lasers also swapped to
use the damage type in use at that time.
* Removal of pocket_data for armor, magazines, tools eetc, restoration
of storage as was used in the build preceding nested containers.