Relic Updates, Part 3
* Fixed some weirdness going on with survivor tachankas causing a
skippable load error.
* Converted mute boots to finally use relic data now that one of the
methods to get it working is functional, and unobsoleted it.
* Converted radiogenic goggles to use mundane radioactivity and light
for part of its relic effects, and implemented its clairvoyance via
relic properties.
* Converted the biological weapons to use relic data. Now that we can
properly cite the effects of mutations with enchantments (i.e. what let
me fix the mute boots), now the flesh weapons use the Degeneration
series instead of the kinda-pointless effect of AEP_SICK, properly give
Schizophrenic while wielded, and give other trait effects based on their
purpose.
* Converted the biological sword to be pure-passive. Since relic data
can't change on transform, the damage buff is pretty much the only thing
you'd get out of tranforming it, and I'm not yet 100% sure how to mock
up "casts a harmful effect to recharge" yet since I only saw JSON
examples for how to generate relics with that feature, no JSON examples
of it on an preexisting item. The upsides and downsides are both
honestly quite potent now, as it is.
* Updated damage type for Hemocatalyst Damage spell, as it seems now
they need to specify damage.