Undo update to mod_tileset, remove from DDA version
So it turns out that the desired method of having all `tiles` entries in
a single `mod_tileset` block behaves in very terrible ways when the mod
tileset needs to cite multiple image files with different sprite sizes
and offsets to behave sanely.

Undid the change to the BN version of the mod tileset and removed the
DDA version for now. I'd reccomend waiting until
https://github.com/CleverRaven/Cataclysm-DDA/issues/46305 is fixed
before re-adding mod tileset to the DDA version, because this seems to
pretty much NEED the ability to define a different mod tileset entity
for each different image file used, at least for the sprite assignments
to not be a giant headache.