Fix weird issue with flamethowers
Turns out that the survivor's flamethrowers were treating the reload time in the same way that magazines treat reload time, despite having an internal magazine. So instead of taking a fixed 20 seconds to reload the whole thing at once, its base reload rate was 20 seconds per single unit of fuel loaded, and with it holding 1500 units of fuel...

Fixed it to be only slightly slower than reloading a vanilla pressurized fuel tank, though combined with its lower capacity it'll still work out to a comparable reload rate (and retaining the disadvantage of being unable to simply swap tanks).