Fix survivor ammo belts in DDA version
I forgot how it's supposed to define armor_data it seems, in the last fix. Odd since I could've sworn it loaded with no errors, but a comment in the commit mentioned an error cropping up somewhere. This time made sure that it loaded in, spawned in, and could be worn with no issue.