Standardized battery pistol/rifle energy efficiency
On reworking the mi-go beam weapon during the previous PR, I made note to myself of the energy-efficiency of related weapons and determined that the survivor battery weapons could perhaps be made a bit less wasteful.

* Changed battery pistol to deal 5 damage and use 20 charges, making it exactly 25% efficient.
* Changed battery rifle to deal 12 damage and use 50 charges, making it 24% efficient.

This still makes them low-power in exchange for being accesible due to using household batteries.