Fix injection of traits into scenarios
Simple enough fix, realized that Cata++ was injecting traits into scenarios in a way that overrides mod injection of traits if it's lower on the load order.

Tested via loading up a world with both Arcana and Cata++, now the martial art traits from both show up in the proper scenarios such as Ambushed and The Next Summer.