Give survivor gun cart reins instead of controls
Companion update to a BN PR, after confirming that reins and tackle actually work as vehicle controls for horse-drawn vehicles.

It does mean you can't cheese the gun cart's turret and remote-firing it by slapping a turret control unit on it, but if you can do that you can already swap the reins for vehicle controls as it is.