Update crossbow item in BN version
If aftermath of https://github.com/CleverRaven/Cataclysm-DDA/pull/21545 ever gets fixed this will need an update to the DDA version too, but for now only BN has fixed it.

Per Noct's suggestion in DMs, their idea for the bolt rifle was an electrically-assisted crossbow so it would count for crossbow gunmods. Also means it gets dampening slot now that that's fixed in BN to actually be usable on crossbows.

Also removed mechanism slot as nothing compatable with crossbows exists currently. Ported over the gunmod slot changes to DDA version too but as mentioned it's been basically 5 years so no telling if the "check for archery skill to screen for crossbows" code will get fixed on their end.