Remove Unstable Genetics from fungal bio-weapon
Turns out having this can give you Carnivore which then makes you unable to eat Mycus fruits.

Potentially I could tinker with improvements to BN's Mycus content in the future too to improve this but is a fast way to correct it plus DDA likely will still run into this.